pyproject.toml
[project]
name = "orchestration-fastapi"
version = "0.1.0"
description = "FastAPI orchestration service"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "python-dotenv==1.0.1",
    "pydantic==2.7.1",
    "watchfiles==1.0.5",
    "websockets==15.0.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py39"
